server/requirements.txt
Flask
Flask-RESTful
Jinja2
mysql-connector